+++ src/main/java/dev/codedsakura/blossom/lib/TeleportUtils.java
@@ -1,24 +1,23 @@
 package dev.codedsakura.blossom.lib;
 
 import net.minecraft.entity.boss.BossBar;
-import net.minecraft.entity.boss.BossBar.Color;
 import net.minecraft.entity.boss.CommandBossBar;
-import net.minecraft.network.packet.s2c.play.ClearTitleS2CPacket;
 import net.minecraft.network.packet.s2c.play.SubtitleS2CPacket;
 import net.minecraft.network.packet.s2c.play.TitleFadeS2CPacket;
 import net.minecraft.network.packet.s2c.play.TitleS2CPacket;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.text.LiteralText;
+import net.minecraft.text.TextColor;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Vec3d;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.UUID;
 
+import static dev.codedsakura.blossom.lib.BlossomLib.CONFIG;
 import static dev.codedsakura.blossom.lib.BlossomLib.LOGGER;
 
 public class TeleportUtils {
@@ -36,16 +35,23 @@ public class TeleportUtils {
         assert server != null;
         final Vec3d[] lastPos = {who.getPos()};
 
-        final TeleportConfig config = customConfig == null ? BlossomLib.CONFIG.baseTeleportation : customConfig.mergeWithDefaults();
+        final TeleportConfig config = customConfig == null ? BlossomLib.CONFIG.baseTeleportation : customConfig.cloneMerge();
 
         CommandBossBar commandBossBar = null;
+        int standTicks = (int) (standStillTime * 20);
         if (config.bossBar.enabled) {
-            commandBossBar = server.getBossBarManager().add(new Identifier(IDENTIFIER), new TranslatableText(""));
-            commandBossBar.addPlayer(who);
+            commandBossBar = server.getBossBarManager().add(
+                    new Identifier(IDENTIFIER + "_" + who.getUuidAsString()),
+                    new TranslatableText(
+                            "blossom.countdown.boss_bar.name",
+                            new LiteralText(Integer.toString(standTicks))
+                                    .styled(style -> style.withColor(TextColor.parse(CONFIG.colors.variable)))
+                    ).styled(style -> style.withColor(TextColor.parse(config.bossBar.textColor)))
+            );
             commandBossBar.setColor(BossBar.Color.byName(config.bossBar.color));
+            commandBossBar.addPlayer(who);
         }
         who.networkHandler.sendPacket(new TitleFadeS2CPacket(0, 10, 5));
-        int standTicks = (int) (standStillTime * 20);
 
         CommandBossBar finalCommandBossBar = commandBossBar;
         TASKS.add(new CounterRunnable(standTicks, who.getUuid()) {
@@ -57,18 +63,21 @@ public class TeleportUtils {
                         finalCommandBossBar.removePlayer(who);
                         server.getBossBarManager().remove(finalCommandBossBar);
                     }
-                    if (config.titleMessageEnabled) {
-                        new Timer().schedule(new TimerTask() {
-                            @Override
-                            public void run() {
-                                who.networkHandler.sendPacket(new ClearTitleS2CPacket(true));
-                            }
-                        }, 500);
+                    if (config.titleMessage.enabled) {
+                        who.networkHandler.sendPacket(new SubtitleS2CPacket(
+                                config.titleMessage.subtitleDone.getText("blossom.countdown.title.done.subtitle")
+                        ));
+                        who.networkHandler.sendPacket(new TitleS2CPacket(
+                                config.titleMessage.titleDone.getText("blossom.countdown.title.done.title")
+                        ));
                     }
 
                     if (counter == 0) {
                         if (config.actionBarMessageEnabled) {
-                            who.sendMessage(new TranslatableText(""), true);
+                            who.sendMessage(
+                                    TextUtils.translation("blossom.countdown.action_bar.done"),
+                                    true
+                            );
                         }
                         onDone.run();
                     }
@@ -88,15 +97,31 @@ public class TeleportUtils {
                     counter = standTicks;
                 }
 
+                int remaining = (int) Math.floor((counter / 20f) + 1);
+
                 if (finalCommandBossBar != null) {
                     finalCommandBossBar.setPercent((float) counter / standTicks);
+                    finalCommandBossBar.setName(
+                            new TranslatableText(
+                                    "blossom.countdown.boss_bar.name",
+                                    new LiteralText(Integer.toString(remaining))
+                                            .styled(style -> style.withColor(TextColor.parse(CONFIG.colors.variable)))
+                            ).styled(style -> style.withColor(TextColor.parse(config.bossBar.textColor)))
+                    );
                 }
                 if (config.actionBarMessageEnabled) {
-                    who.sendMessage(new TranslatableText(""), true);
+                    who.sendMessage(
+                            TextUtils.translation("blossom.countdown.action_bar.counting", remaining),
+                            true
+                    );
                 }
-                if (config.titleMessageEnabled) {
-                    who.networkHandler.sendPacket(new SubtitleS2CPacket(new TranslatableText("")));
-                    who.networkHandler.sendPacket(new TitleS2CPacket(new TranslatableText("")));
+                if (config.titleMessage.enabled) {
+                    who.networkHandler.sendPacket(new SubtitleS2CPacket(
+                            config.titleMessage.subtitleCounting.getText("blossom.countdown.title.counting.subtitle", remaining)
+                    ));
+                    who.networkHandler.sendPacket(new TitleS2CPacket(
+                            config.titleMessage.titleCounting.getText("blossom.countdown.title.counting.title", remaining)
+                    ));
                 }
             }
         });
@@ -129,45 +154,6 @@ public class TeleportUtils {
 
         boolean shouldRemove() {
             return counter < 0;
-        }
-    }
-
-    public static class TeleportConfig {
-        public TeleportConfig(boolean defaults) {
-            bossBar = defaults ? new BossBar(true) : null;
-        }
-
-        BossBar bossBar;
-
-        static class BossBar {
-            boolean enabled = true;
-            String color;
-
-            public BossBar(boolean defaults) {
-                color = defaults ? Color.PINK.getName() : null;
-            }
-        }
-
-        boolean actionBarMessageEnabled = false;
-
-        boolean titleMessageEnabled = true;
-
-        boolean fovEffectEnabled = false;
-
-        ParticleAnimation particleAnimation = ParticleAnimation.OFF;
-
-        enum ParticleAnimation {
-            OFF
-        }
-
-        TeleportConfig mergeWithDefaults() {
-            TeleportConfig defaultConf = BlossomLib.CONFIG.baseTeleportation;
-
-            TeleportConfig newOne = new TeleportConfig(false);
-            newOne.bossBar = this.bossBar == null ? defaultConf.bossBar : this.bossBar;
-            newOne.bossBar.color = newOne.bossBar.color == null ? defaultConf.bossBar.color : newOne.bossBar.color;
-
-            return newOne;
         }
     }
 }